requirements.txt
selenium-wire==5.1.0
jmespath==1.0.1
selenium~=4.7.2
cerberus==1.3.4
yoyo-migrations==8.1.0
pytest==7.2.0
Flask==2.2.2
tweepy==4.12.1